Move StateEffect to symbols
